refactor: update files to typescript
