--- src/users.js
@@ -1,11 +1,0 @@
-const users = [];
-
-const testUser = {
-  id: "001",
-  username: "Test user",
-  age: 18,
-  hobbies: "tennis",
-};
-users.push(testUser);
-
-export { users };
